--- src/website_update/app.py
@@ -1,28 +1,0 @@
-from flask import Flask, render_template, request, jsonify
-import os
-
-app = Flask(__name__)
-
-@app.route('/')
-def index():
-    return render_template('index.html')
-
-@app.route('/<image_id>')
-def get_image_url(image_id):
-    # Implement logic to fetch the image URL based on the image_id
-    # This could involve querying a database or some other storage mechanism
-    # Return the image URL in the format {"url": <url>}
-    return jsonify({"url": "https://example.com/path/to/image.jpg"})
-
-@app.route('/submit', methods=['POST'])
-def submit_form():
-    transform = request.form['transform']
-    # Process the image file and obtain the image_id
-    # This could involve saving the file, applying the transformation, and storing metadata
-    image_id = "123456"  # Replace with the actual image_id
-
-    # Return the image_id in the format {"image_id": <image_id>}
-    return jsonify({"image_id": image_id})
-
-if __name__ == '__main__':
-    app.run(debug=True)
